+++ plantcv/geospatial/visualize/__init__.py
@@ -1,0 +1,4 @@
+from plantcv.geospatial.visualize.height import height_distribution
+
+
+__all__ = ["height_distribution"]